Make Configure(Web)?HostBuilder public
- simplify sample
