+++ src/Http/samples/MinimalSample/Program.cs
@@ -1,10 +1,9 @@
 using System;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Mvc;
 
 await using var app = WebApplication.Create();
 
-Todo EchoTodo([FromBody] Todo todo) => todo;
+Todo EchoTodo(Todo todo) => todo;
 app.MapPost("/EchoTodo", (Func<Todo, Todo>)EchoTodo);
 
 string Plaintext() => "Hello, World!";
